Moved make tape modal to its own file and component
