+++ client/components/index.js
@@ -8,5 +8,6 @@ export {default as UserHome} from './user-home';
 export {default as Player} from './player';
 export {default as UploadSongs} from './upload-songs';
 export {default as MyTapes} from './my-tapes';
+export {default as MakeTape} from './make-tape';
 export {default as EditTape} from './edit-tape';
 export {Login, Signup} from './auth-form';
